+++ src/App.jsx
@@ -1,7 +1,8 @@
+import { Routes, Route } from "react-router-dom";
+
 import "./App.css";
 import { Header } from "./Componnets/header/Header";
 import { Trending } from "./pages/trending/Trending";
-import { Routes, Route } from "react-router-dom";
 import { Footer } from "./footer/footer";
 
 function App() {
