+++ src/Componnets/movie-card/MovieCard.css
@@ -26,7 +26,7 @@
 .movie-date,
 .movie-type {
     font-size: 16px;
-    color: rgb(128, 128, 128);
+    color: #808080;
     padding: 0 3px;
     font-weight: 500;
 }
@@ -51,7 +51,7 @@
 }
 
 .high-rating {
-    background-color: rgb(0, 0, 138);
+    background-color: #00008a;
 }
 
 .movie-card:hover>.movie-title {
