+++ src/Componnets/movie-card/MovieCard.jsx
@@ -1,22 +1,25 @@
 import React from "react";
-import { Poster } from "../poster/Poster";
+
 import "./MovieCard.css";
+import { Poster } from "../poster/Poster";
 
 export const MovieCard = ({
   name,
   posterPath,
   title,
-  voteAverage = 5,
+  voteAverage,
   mediaType,
   releaseDate,
 }) => {
   return (
     <div className="movie-card">
-      <span
-        className={`rate ${voteAverage < 6 ? "low-rating" : "high-rating"}`}
-      >
-        {voteAverage}
-      </span>
+      {!!voteAverage && (
+        <span
+          className={`rate ${voteAverage < 6 ? "low-rating" : "high-rating"}`}
+        >
+          {voteAverage}
+        </span>
+      )}
       <Poster name={name} src={posterPath} />
       <h2 className="movie-title">{title ?? name}</h2>
       <div className="movie-footer">
