+++ src/Componnets/pagination/pagination.jsx
@@ -1,5 +1,7 @@
 import React from "react";
+
 import "./pagination.css";
+import { getPagesTodisplay } from "../../util/pagination";
 
 export const Pagination = ({
   countPage,
@@ -7,31 +9,11 @@ export const Pagination = ({
   handlePageClick,
   currentPage,
 }) => {
-  const pagesList = [...Array(countPage + 1).keys()].slice(1);
-
-  const pagesTodisplay = pagesList
-    .filter((page) => {
-      if (page === 1 || page === pagesList.length) return true;
-      if (
-        page >= currentPage - 1 &&
-        page < currentPage + marginPagesDisplayed
-      ) {
-        return true;
-      }
-      return false;
-    })
-    .reduce((pages, page, index, pagesList) => {
-      if (index === 1 && page - 1 !== 1) {
-        pages.push("...");
-      }
-
-      pages.push(page);
-
-      if (index === pagesList.length - 2 && page + 1 !== countPage) {
-        pages.push("...");
-      }
-      return pages;
-    }, []);
+  const pagesTodisplay = getPagesTodisplay(
+    countPage,
+    marginPagesDisplayed,
+    currentPage
+  );
 
   return (
     <ul className="pagination">
